requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
azure-ai-formrecognizer>=3.3.3
azure-search-documents>=11.5.2
azure-identity>=1.21.0
openai>=1.74.0
PyPDF2>=3.0.0